requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
discord-protos==0.0.2
discord.py-self @ git+https://github.com/dolfies/discord.py-self.git@b813c625dcaee7cd013ad28f7ae348339f4e02a4
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
protobuf==5.27.0
python-dotenv==1.0.1
tzlocal==5.2
yarl==1.9.4
